[Fixed Postion] Closed
Currently the Fixed controller will add the
element created to fixed position view, therefore
to the elementProxies list and the newly made elementConfigByIds
list (made from .configuration.elements). This will add
a default position based on the index value, however when deleted
using edit toolbar, the element is not actually deleted from the
fixed position view and will reappear. If deleted from the element
list, it will actually delete, but will create an undefined ghost
of the element in the fixed position view that is not seen in
browse mode, but has an outline in edit mode's fixed position view.
WTD 17.